requirements.txt
msgpack-python==0.5.6
Pillow==8.2.0
pkg-resources==0.0.0
pyfinite==1.7
